Adding unit test in RestaurantServiceTest Also included a verification if the restaurant exists in db otherwise it'll throw an exception

+++ src/main/java/com/example/bootiful_restaurant_reviews/service/RestaurantService.java
@@ -34,6 +34,8 @@ public class RestaurantService {
     }
 
     public void deleteRestaurantById(Long id) {
-        restaurantRepository.deleteById(id);
+        if (restaurantRepository.existsById(id)) {
+            restaurantRepository.deleteById(id);
+        }
     }
 }
